+++ server/src-lib/Hasura/RQL/Types.hs
@@ -59,6 +59,7 @@ import           Hasura.RQL.Types.GraphqlSchemaIntrospection as R
 import           Hasura.RQL.Types.Metadata                   as R
 import           Hasura.RQL.Types.Metadata.Backend           as R
 import           Hasura.RQL.Types.Metadata.Object            as R
+import           Hasura.RQL.Types.Network                    as R
 import           Hasura.RQL.Types.Permission                 as R
 import           Hasura.RQL.Types.QueryCollection            as R
 import           Hasura.RQL.Types.QueryTags                  as R