+++ server/src-lib/Hasura/Server/Migrate.hs
@@ -281,7 +281,7 @@ migrations maybeDefaultSourceConfig dryRun maintenanceMode =
                     SourceMetadata defaultSource _mnsTables _mnsFunctions defaultSourceConfig
               in Metadata (OMap.singleton defaultSource defaultSourceMetadata)
                    _mnsRemoteSchemas _mnsQueryCollections _mnsAllowlist _mnsCustomTypes _mnsActions _mnsCronTriggers mempty
-                   emptyApiLimit emptyMetricsConfig mempty mempty emptyQueryTagsConfig
+                   emptyApiLimit emptyMetricsConfig mempty mempty emptyQueryTagsConfig emptyNetwork
         liftTx $ insertMetadataInCatalog metadataV3
 
     from43To42 = do